Add tests for unlimited
